+++ src/index.ts
@@ -430,14 +430,16 @@ export {
 
 export {
     buffer,
+    cookies,
     defaultBodyLimit,
     defaultLimitReachedHandler,
     defaultParseErrorHandler,
     defaultValidationFailedHandler,
-    json,
     IBufferOptions,
     IJsonOptions,
     IValidateOptions,
+    json,
+    lang,
     query,
     validate
 } from './middlewares';