+++ src/middlewares/buffer.ts
@@ -46,13 +46,13 @@ interface ICreateMiddlewareOptions {
  * const app = createServer()
  *
  * async function handleLimitReached(request: IHttpRequest, response: IHttpResponse) {
- *     request.writeHead(400)
- *     request.write('Input is too big')
+ *   request.writeHead(400)
+ *   request.write('Input is too big')
  * }
  *
  * // maximum input size: 128 MB
  * app.post('/', buffer(), async (request: IHttpRequest, response: IHttpResponse) => {
- *     assert.strictEqual(Buffer.isBuffer(request.body), true)
+ *   assert.strictEqual(Buffer.isBuffer(request.body), true)
  * })
  *
  * // maximum input size: 256 MB
@@ -62,14 +62,14 @@ interface ICreateMiddlewareOptions {
  *
  * // maximum input size: 384 MB
  * app.patch('/', buffer({ limit: 402653184 }), async (request: IHttpRequest, response: IHttpResponse) => {
- *     assert.strictEqual(Buffer.isBuffer(request.body), true)
+ *   assert.strictEqual(Buffer.isBuffer(request.body), true)
  * })
  *
  * // custom error handler
  * app.delete('/', buffer({ limit: 1048576, onLimitReached: handleLimitReached }), async (request: IHttpRequest, response: IHttpResponse) => {
  * // alternative:
  * // app.delete('/', buffer(1, handleLimitReached), async (request: IHttpRequest, response: IHttpResponse) => {
- *     assert.strictEqual(Buffer.isBuffer(request.body), true)
+ *   assert.strictEqual(Buffer.isBuffer(request.body), true)
  * })
  * ```
  */